Constants

Constants

Many PETSc constants and enums are available in the Julia wrappers. The full list can be found in src/petsc_constants.jl. See the PETSc documentation for their meaning. A few of the most important constants are listed here. Not all of the constants are exported from the PETSc module, to avoid cluttering the user's namespace.

PETSc2.PetscInt — Type.

Petsc integer type

source

PETSc2.PetscReal — Type.

The closest real type to PetscScalar

source

PETSc2.PetscScalar — Type.

Element type of Petsc vectors and matrices

source