Error Handling Documentation

This page describes error handling in the Julia wrappers.

PETSc.error_handler — Method.

Error handler registered with PETSc. This function prints the information Petsc supplies about the error, a Julia stack trace, and then invokes the PetscTraceBackErrorHandler to print the Petsc stack trace.