Mat Documentation

This page wraps functions in PETSc's Mat API. Consult the PETSc documentation for the behavior of the functions. The documentation on this page only describes differences between the Julia wrappers and the PETSc documentation.

PETSc.PetscMat — Type.

PetscMat type. Currently a subtype of AbstractArray, although that may change.

PETSc.MatAXPY — Method.

MatAxPY

PETSc.MatAYPX — Method.

MatAYPX

PETSc.MatAssemblyBegin — Method.

MatAssemblyBegin

PETSc.MatAssemblyBegin — Method.

MatAssemblyBegin, impicitly using PETSC_MAT_FINAL_ASSEMBLY

PETSc.MatAssemblyEnd — Method.

MatAssemblyEnd

PETSc.MatAssemblyEnd — Method.

MatAssemblyEnd, implicitly using PETSC_MAT_FINAL_ASSEMBLY

PETSc.MatCreateShell — Method.

MatCreateShell

PETSc.MatCreateTranspose — Method.

MatCreateTranspose

PETSc.MatGetInfo — Method.

MatGetInfo

PETSc.MatGetLocalSize — Method.

MatGetLocalsize. Same interface as MatGetSize

PETSc.MatGetOwnershipRange — Method.

MatGetOwnershipRange

Inputs

mat: a PetscMat

Outputs

low: lowest (zero-based) index that is owned

high: highest + 1 (zero-based) index that is owned

PETSc.MatGetSize — Method.

MatGetSize

Inputs

obj: a PetscMat

Outputs

m: first dimension size

n: second dimension size

PETSc.MatGetType — Method.

MatGetType

PETSc.MatGetValues — Method.

MatGetValues

Inputs

obj: the PetscMat

idxm: row indices

idxn: column indices

vals: logically 2 dimensional array of values (although it could be a vector too)

PETSc.MatMPIAIJSetPreallocation — Method.

MatMPIAIJSetPreallocation

PETSc.MatMatMult — Method.

MatMatMult

PETSc.MatMult — Method.

MatMult

PETSc.MatMultAdd — Method.

MatMultAdd

PETSc.MatMultHermitianTranspose — Method.

MatMultHermitianTranspose

PETSc.MatMultTranspose — Method.

MatMultTranspose

PETSc.MatNorm — Method.

MatNorm

PETSc.MatScale — Method.

MatScale

PETSc.MatSetFromOptions — Method.

MatSetFromOptions

PETSc.MatSetOption — Method.

MatSetOption

PETSc.MatSetSizes — Method.

MatSetSizes

PETSc.MatSetType — Method.

MatSetType

PETSc.MatSetValues — Method.

MatSetValues. The idx and idy arrays must have PetscInt elements, and the vals array must have PetscScalar elements.

PETSc.MatSetValuesBlocked — Method.

MatSetValues blocked.

PETSc.MatShellGetContext — Method.

MatShellGetContext

PETSc.MatShellSetOperation — Method.

MatShellSetOperation

PETSc.MatShift — Method.

MatShift

PETSc.MatTranspose — Method.

Constructs the transpose of a given matrix. Can be in place or out of place.

Inputs

A: matrix to take the transpose of

Keywords

inplace: whether or not to do the transpose in place

Outputs

A PetscMat object, either the original object A if the transpose was done in place, or a new matrix object if the transpose was done out of place

PETSc.MatXAIJSetPreallocation — Method.

MatXAIJSetPreallocation

PETSc.MatZeroEntries — Method.

MatZeroEntries

PETSc.PetscDestroy — Method.

Frees a Petsc object. Safe to call multiple times

PETSc.PetscView — Function.

PetscViewer

PETSc.SetUp — Method.

MatSetUp

PETSc.MatInfo — Type.

Equivalent to Petsc's MatInfo struct

Base.show — Method.

show() for a MatInfo