Vec Documentation

Vec Documentation

This page wraps functions in PETSc's Vec API. Consult the PETSc documentation for the behavior of the functions. The documentation on this page only describes differences between the Julia wrappers and the PETSc documentation.

PETSc2.PetscVec — Type.

Petsc Vector type.

Not a subtype of AbstractArray because Petsc vectors do not conform to that API.

source

PETSc2.PetscDestroy — Method.

Free a Petsc vec. Safe to call multiple times

source

PETSc2.PetscView — Method.

PetscView for Petsc vector

source

PETSc2.VecAXPBY — Method.

VecAXPBY

source

PETSc2.VecAXPBYPCZ — Method.

VecAXPBYPCZ

source

PETSc2.VecAXPY — Method.

VecAXPY

source

PETSc2.VecAYPX — Method.

VecAYPX

source

PETSc2.VecAssemblyBegin — Method.

VecAssemblyBegin

source

PETSc2.VecAssemblyEnd — Method.

VecAssemblyEnd

source

PETSc2.VecCopy — Method.

VecCopy

source

PETSc2.VecDot — Method.

VecDot

source

PETSc2.VecDuplicate — Method.

VecDulicate

source

PETSc2.VecExp — Method.

VecExp

source

PETSc2.VecGetArray — Method.

VecGetArray. Users must call VecRestoreArray when finished.

Inputs

vec: the Petsc vector

Outputs

arr: a Julia Array{PetscScalar, 1}

source

PETSc2.VecGetArrayRead — Method.

Similar to VecGetArray, but produces the array must not be written to.

source

PETSc2.VecGetLocalSize — Method.

VecGetLocalSize

Inputs

vec: a Petsc vector

Outputs

the size

source

PETSc2.VecGetOwnershipRange — Method.

VecGetOwnershipRange

Inputs

vec: Petsc vector

Outputs

low: lowest index (zero-based) that is owned

high: highest index + 1 that is owned

source

PETSc2.VecGetSize — Method.

VeGetSize

Inputs

vec: a Petsc vector

Outputs

the size

source

PETSc2.VecGetValues — Method.

VecGetValues with length of idx inferred from idx

Inputs

vec: the Petsc vector

idx: array of PetscInt indices

y: array of PetscScalar values

source

PETSc2.VecGetValues — Method.

VecGetValues

source

PETSc2.VecLog — Method.

VecLog

source

PETSc2.VecMAXPY — Method.

VecMAXPY

source

PETSc2.VecMax — Method.

VecMax

Inputs

vec: PetscVec

Output

r: the maximum value

idx: the (zero-based) index of the maximum value

source

PETSc2.VecMin — Method.

VecMin. Same interface as VecMax

source

PETSc2.VecNorm — Method.

VecNorm

Inputs

obj: Petsc vector

normtype: the Petsc enum for the norm type

Output

the norm value

source

PETSc2.VecPointwiseDivide — Method.

VecPointwiseDivide

source

PETSc2.VecPointwiseMult — Method.

VecPointwiseMult

source

PETSc2.VecReciprocal — Method.

VecReciprocal

source

PETSc2.VecRestoreArray — Method.

VecRestoreArray. Users must not access the array after calling this function.

Inputs

vec: the PetscVector passed into VecGetArray

arr: the array returned by `VecGetArray

source

PETSc2.VecRestoreArrayRead — Method.

Similar to VecRestoreArray, but corresponds to VecGetArrayRead

source

PETSc2.VecScale — Method.

VecScale

source

PETSc2.VecSet — Method.

VecSet

source

PETSc2.VecSetSizes — Method.

VecSetSizes

source

PETSc2.VecSetType — Method.

VecSetType

source

PETSc2.VecSetValues — Method.

VecSetValues

source

PETSc2.VecSetValues — Method.

VecSetValues method that implicitly uses INSERT_VALUES

source

PETSc2.VecShift — Method.

VecShift

source

PETSc2.VecSqrtAbs — Method.

VecSqrtAbs

source

PETSc2.VecSum — Method.

VecSum

source

PETSc2.VecSwap — Method.

VecSwap

source

PETSc2.VecTDot — Method.

VecTDot

source

PETSc2.VecWAXPY — Method.

VecWAXPY

source

PETSc2.AllVectors — Constant.

Union{AbstractVector, PetscVec}

source

PETSc2.VecAssemble — Method.

Convenience function for calling VecAssemblyBegin, and VecAssemblyEnd, in one go.

source