Mat Documentation

Mat Documentation

This page wraps functions in PETSc's Mat API. Consult the PETSc documentation for the behavior of the functions. The documentation on this page only describes differences between the Julia wrappers and the PETSc documentation.

PETSc2.PetscMat — Type.

PetscMat type. Currently a subtype of AbstractArray, although that may change.

source

PETSc2.MatAXPY — Method.

MatAxPY

source

PETSc2.MatAYPX — Method.

MatAYPX

source

PETSc2.MatAssemblyBegin — Method.

MatAssemblyBegin

source

PETSc2.MatAssemblyBegin — Method.

MatAssemblyBegin, impicitly using MAT_FINAL_ASSEMBLY

source

PETSc2.MatAssemblyEnd — Method.

MatAssemblyEnd

source

PETSc2.MatAssemblyEnd — Method.

MatAssemblyEnd, implicitly using MAT_FINAL_ASSEMBLY

source

PETSc2.MatCreateShell — Method.

MatCreateShell

source

PETSc2.MatCreateTranspose — Method.

MatCreateTranspose

source

PETSc2.MatGetInfo — Method.

MatGetInfo

source

PETSc2.MatGetLocalSize — Method.

MatGetLocalsize. Same interface as MatGetSize

source

PETSc2.MatGetOwnershipRange — Method.

MatGetOwnershipRange

Inputs

mat: a PetscMat

Outputs

low: lowest (zero-based) index that is owned

high: highest + 1 (zero-based) index that is owned

source

PETSc2.MatGetSize — Method.

MatGetSize

Inputs

obj: a PetscMat

Outputs

m: first dimension size

n: second dimension size

source

PETSc2.MatGetType — Method.

MatGetType

source

PETSc2.MatGetValues — Method.

MatGetValues

Inputs

obj: the PetscMat

idxm: row indices

idxn: column indices

vals: logically 2 dimensional array of values (although it could be a vector too)

source

PETSc2.MatMPIAIJSetPreallocation — Method.

MatMPIAIJSetPreallocation

source

PETSc2.MatMatMult — Method.

MatMatMult

source

PETSc2.MatMult — Method.

MatMult

source

PETSc2.MatMultAdd — Method.

MatMultAdd

source

PETSc2.MatMultHermitianTranspose — Method.

MatMultHermitianTranspose

source

PETSc2.MatMultTranspose — Method.

MatMultTranspose

source

PETSc2.MatNorm — Method.

MatNorm

source

PETSc2.MatScale — Method.

MatScale

source

PETSc2.MatSetOption — Method.

MatSetOption

source

PETSc2.MatSetSizes — Method.

MatSetSizes

source

PETSc2.MatSetType — Method.

MatSetType

source

PETSc2.MatSetValues — Method.

MatSetValues. The idx and idy arrays must have PetscInt elements, and the vals array must have PetscScalar elements.

source

PETSc2.MatSetValuesBlocked — Method.

MatSetValues blocked.

source

PETSc2.MatShellGetContext — Method.

MatShellGetContext

source

PETSc2.MatShellSetOperation — Method.

MatShellSetOperation

source

PETSc2.MatShift — Method.

MatShift

source

PETSc2.MatTranspose — Method.

Constructs the transpose of a given matrix. Can be in place or out of place.

Inputs

A: matrix to take the transpose of

Keywords

inplace: whether or not to do the transpose in place

Outputs

A PetscMat object, either the original object A if the transpose was done in place, or a new matrix object if the transpose was done out of place

source

PETSc2.MatXAIJSetPreallocation — Method.

MatXAIJSetPreallocation

source

PETSc2.MatZeroEntries — Method.

MatZeroEntries

source

PETSc2.PetscDestroy — Method.

Frees a Petsc object. Safe to call multiple times

source

PETSc2.PetscView — Function.

PetscViewer

source

PETSc2.SetFromOptions — Method.

MatSetFromOptions

source

PETSc2.SetUp — Method.

MatSetUp

source

PETSc2.MatInfo — Type.

Equivalent to Petsc's MatInfo struct

source

Base.show — Method.

show() for a MatInfo

source