Options Documentation

This page wraps functions in PETSc's options API. Consult the PETSc documentation for the behavior of the functions. The documentation on this page only describes differences between the Julia wrappers and the PETSc documentation.

PETSc.PetscClearOptions — Method.

Convenience wrapper for using a dictionary to clear options (only the keys are used).

PETSc.PetscOptionsClearValue — Function.

PetscOptionsClearValue

Inputs

arg1: the key (string)

arg2: the PetscOptions object, defaults to the global options databse

PETSc.PetscOptionsSetValue — Function.

PetscOptionsSetValue

Inputs

arg1: the key (string)

arg2: the value (string)

arg3: the PetscOptions object, defaults to the global options database

PETSc.PetscOptionsView — Function.

PetscOptionsView

Inputs

arg1: a PetscViewer, defaults to Petsc's stdout

arg2: the PetscOptions object, defaults to the global options databse

PETSc.PetscSetOptions — Method.

Convenience wrapper for using a dictionary to set options

PETSc.PetscOptions — Type.

Typedef of PetscOptions