Vec Documentation

This page wraps functions in PETSc's Vec API. Consult the PETSc documentation for the behavior of the functions. The documentation on this page only describes differences between the Julia wrappers and the PETSc documentation.

PETSc.PetscVec — Type.

Petsc Vector type.

Not a subtype of AbstractArray because Petsc vectors do not conform to that API.

PETSc.PetscDestroy — Method.

Free a Petsc vec. Safe to call multiple times

PETSc.PetscView — Method.

PetscView for Petsc vector

PETSc.VecAXPBY — Method.

VecAXPBY

PETSc.VecAXPBYPCZ — Method.

VecAXPBYPCZ

PETSc.VecAXPY — Method.

VecAXPY

PETSc.VecAYPX — Method.

VecAYPX

PETSc.VecAssemblyBegin — Method.

VecAssemblyBegin

PETSc.VecAssemblyEnd — Method.

VecAssemblyEnd

PETSc.VecCopy — Method.

VecCopy

PETSc.VecDot — Method.

VecDot

PETSc.VecDuplicate — Method.

VecDulicate

PETSc.VecExp — Method.

VecExp

PETSc.VecGetArray — Method.

VecGetArray. Users must call VecRestoreArray when finished.

Inputs

vec: the Petsc vector

Outputs

arr: a Julia Array{PetscScalar, 1}

PETSc.VecGetArrayRead — Method.

Similar to VecGetArray, but produces the array must not be written to.

PETSc.VecGetLocalSize — Method.

VecGetLocalSize

Inputs

vec: a Petsc vector

Outputs

the size

PETSc.VecGetOwnershipRange — Method.

VecGetOwnershipRange

Inputs

vec: Petsc vector

Outputs

low: lowest index (zero-based) that is owned

high: highest index + 1 that is owned

PETSc.VecGetSize — Method.

VeGetSize

Inputs

vec: a Petsc vector

Outputs

the size

PETSc.VecGetValues — Method.

VecGetValues with length of idx inferred from idx

Inputs

vec: the Petsc vector

idx: array of PetscInt indices

y: array of PetscScalar values

PETSc.VecGetValues — Method.

VecGetValues

PETSc.VecLog — Method.

VecLog

PETSc.VecMAXPY — Method.

VecMAXPY

PETSc.VecMax — Method.

VecMax

Inputs

vec: PetscVec

Output

r: the maximum value

idx: the (zero-based) index of the maximum value

PETSc.VecMin — Method.

VecMin. Same interface as VecMax

PETSc.VecNorm — Method.

VecNorm

Inputs

obj: Petsc vector

normtype: the Petsc enum for the norm type

Output

the norm value

PETSc.VecPointwiseDivide — Method.

VecPointwiseDivide

PETSc.VecPointwiseMult — Method.

VecPointwiseMult

PETSc.VecReciprocal — Method.

VecReciprocal

PETSc.VecRestoreArray — Method.

VecRestoreArray. Users must not access the array after calling this function.

Inputs

vec: the PetscVector passed into VecGetArray

arr: the array returned by `VecGetArray

PETSc.VecRestoreArrayRead — Method.

Similar to VecRestoreArray, but corresponds to VecGetArrayRead

PETSc.VecScale — Method.

VecScale

PETSc.VecSet — Method.

VecSet

PETSc.VecSetSizes — Method.

VecSetSizes

PETSc.VecSetType — Method.

VecSetType

PETSc.VecSetValues — Method.

VecSetValues

PETSc.VecSetValues — Method.

VecSetValues method that implicitly uses PETSC_INSERT_VALUES

PETSc.VecShift — Method.

VecShift

PETSc.VecSqrtAbs — Method.

VecSqrtAbs

PETSc.VecSum — Method.

VecSum

PETSc.VecSwap — Method.

VecSwap

PETSc.VecTDot — Method.

VecTDot

PETSc.VecWAXPY — Method.

VecWAXPY

PETSc.AllVectors — Constant.

Union{AbstractVector, PetscVec}

PETSc.VecAssemble — Method.

Convenience function for calling VecAssemblyBegin, and VecAssemblyEnd, in one go.